[Tag] refactoring and creating inMemory controller
