+++ 04-delivery/src/test/scala/com/khanr1/cryocompose/controllers/TagControllerSuite.scala
@@ -1,0 +1,5 @@
+package com.khanr1.cryocompose.controllers
+
+object TagControllerSuite {
+  
+}
